requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
idna==2.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycryptodome==3.23.0
PyJWT==2.10.1
pytz==2025.2
requests==2.18.4
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==1.22
Werkzeug==3.1.3
